chore(ci): explicitly trigger PR workflows after syncing specs
This updates the existing spec sync jobs and the sync job template
to "manually" trigger the PR validation workflows after the spec
sync PR is opened.

The PR validation workflows have been updated to listen for the
`workflow_dispatch` event; if the workflows were triggered by this
event and the actor was GitHub Actions, the workflows automatically
set the relevant commit status so that we can enable required status
checks for all PRs.